fix: sun icon feedback (#124) + quick agent-command menu on My Day (#152)
#124 — Backlog sun icon now reflects My Day state

The sun-icon "Toggle My Day" button on Backlog list-card and board-card
rows had no Foreground binding to task.IsMyDay. Tasks toggled correctly
behind the scenes (frontmatter wrote, vault was right) but the UI gave
no visual confirmation. The same MyDayBrushConverter pattern was already
in use for TaskDetail subtask suns (sunny gold when active, muted gray
when inactive) — but the converter lived inside TaskDetailPage's
code-behind in the local: namespace, so other pages couldn't reach it.

Lifted MyDayBrushConverter to a new file under Glasswork.Converters
(alongside RowConverters), updated TaskDetailPage to reference it via
the conv: namespace, declared it in BacklogPage.xaml's resources, added
Foreground bindings on both Backlog sun buttons (list-card line 54 and
board-card line 180). My Day card uses a separate "Remove from My Day"
X-icon affordance, not a sun toggle, so no change there.

#152 — Quick-copy agent-command menu on My Day cards

User reported having to drill into TaskDetail and scroll past the entire
task body just to copy the Start/Resume/Wrap-up agent-command lines.
Added a ⋯ MenuFlyout button to each My Day card title row (right of the
existing X button) with three items that copy the appropriate
TaskInvocationFormatter output to the clipboard. Feedback surfaces
through a top-right InfoBar styled to match the existing precedent in
TaskDetailPage.

TaskInvocationFormatter (Glasswork.Core.Services) already produces the
canonical strings and is fully unit-tested, so no new domain logic was
needed. The clipboard-set + toast pattern mirrors TaskDetailPage's
ClipboardHint to stay consistent across surfaces.

Verification

* dotnet build src/Glasswork.App -r win-x64 → 0 errors, 2 pre-existing
  warnings (TaskDetailPage.xaml.cs:375 / :478, out of scope here).
* 495/496 unit tests pass. The single failure is
  DebouncerTests.Trigger_FiresAgainAfterQuietPeriodElapses, an unrelated
  flaky timing-based test in a Core utility — see issue #155 (filed
  separately).
* Manually verified in the freshly-published Release binary:
  - Backlog list view: sun toggle flips between gold and gray on click.
  - Backlog board view: same.
  - TaskDetail subtask suns: still work (regression check).
  - My Day card: ⋯ menu shows three items; each copies the right
    TaskInvocationFormatter line and surfaces a "Copied 'X' command"
    InfoBar.
